RAG_Server/requirements.txt
# Core LangChain
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-huggingface>=0.0.3
langchain-openai>=0.1.0            # OpenRouter uses the OpenAI-compatible API

# Vector store
chromadb>=0.5.0

# Embeddings + Reranker
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.2.0                        # CPU-only: torch --index-url https://download.pytorch.org/whl/cpu

# BM25
rank-bm25>=0.2.2

# PDF loading
PyPDF2>=3.0.0

# News / RSS
feedparser>=6.0.0
newsapi-python>=0.2.7               # optional – for NewsAPI

# Weather / HTTP
httpx>=0.27.0

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# Utilities
numpy>=1.26.0
python-dotenv>=1.0.0